requirements.txt
streamlit~=1.38.0
streamlit-authenticator~=0.3.3
boto3~=1.33.4
opensearch-py==2.4.2
langchain~=0.1.11
langchain-core~=0.1.30
pandas
fastapi~=0.110.1
numpy
python-dotenv